+++ packages/e2e-tests/src/tests/invitationLink.test.ts
@@ -7,7 +7,6 @@ import {
   RegisterUsernameModal,
   App,
   Sidebar,
-  StartingLoadingPanel,
   WarningModal,
 } from '../selectors'
 import { capitalizeFirstLetter, invitationDeepUrl } from '@quiet/common'
@@ -30,6 +29,10 @@ describe('New user joins using invitation link while having app opened', () => {
     guestApp = new App({ useDataDir: false })
   })
 
+  beforeEach(async () => {
+    await new Promise<void>(resolve => setTimeout(() => resolve(), 1000))
+  })
+
   afterAll(async () => {
     await ownerApp?.close()
     await guestApp?.close()
@@ -48,13 +51,6 @@ describe('New user joins using invitation link while having app opened', () => {
         await debugModal.close()
       })
     }
-
-    it('StartingLoadingPanel modal', async () => {
-      console.log('Invitation Link', 3)
-      const loadingPanel = new StartingLoadingPanel(ownerApp.driver)
-      const isLoadingPanel = await loadingPanel.element.isDisplayed()
-      expect(isLoadingPanel).toBeTruthy()
-    })
 
     it('JoinCommunityModal - owner switches to create community', async () => {
       console.log('Invitation Link', 4)