+++ packages/mobile/src/setupTests.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable */
 import { setEngine, CryptoEngine } from 'pkijs'
 import { Crypto } from '@peculiar/webcrypto'
+import React from 'react'
 
 import { io } from 'socket.io-client'
 
@@ -24,7 +25,7 @@ jest.mock('react-native-config', () => ({
   NODE_ENV: 'staging',
 }))
 
-jest.mock('redux-persist-filesystem-storage', () => {})
+jest.mock('redux-persist-filesystem-storage', () => { })
 
 jest.mock('redux-persist', () => {
   const real = jest.requireActual('redux-persist')
@@ -36,7 +37,7 @@ jest.mock('redux-persist', () => {
 
 jest.mock('react-native/Libraries/EventEmitter/NativeEventEmitter')
 
-jest.mock('react-native-mathjax-html-to-svg', () => {})
+jest.mock('react-native-mathjax-html-to-svg', () => { })
 
 jest.mock('react-native-qrcode-svg', () => jest.fn())
 
@@ -44,13 +45,22 @@ jest.mock('react-native-progress', () => ({
   CircleSnail: jest.fn(),
 }))
 
+jest.mock(
+  '@ronradtke/react-native-markdown-display', () => ({
+    __esModule: true,
+    default: (props: any) => {
+      return <div>{props.children}</div>
+    },
+    MarkdownIt: jest.fn()
+  }))
+
 jest.mock('socket.io-client', () => ({
   io: jest.fn(),
 }))
 
 // Mocked because of: "Invariant Violation: TurboModuleRegistry.getEnforcing(...): 'RNDocumentPicker'
 // could not be found. Verify that a module by this name is registered in the native binary."
-jest.mock('react-native-document-picker', () => {})
+jest.mock('react-native-document-picker', () => { })
 
 export const ioMock = io as jest.Mock