+++ packages/identity/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.0](https://github.com/TryQuiet/identity/compare/@quiet/identity@1.9.0-alpha.0...@quiet/identity@2.0.0-alpha.0) (2023-09-01)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [1.9.0-alpha.0](https://github.com/TryQuiet/identity/compare/@quiet/identity@1.8.0...@quiet/identity@1.9.0-alpha.0) (2023-08-29)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [1.8.0](https://github.com/TryQuiet/identity/compare/@quiet/identity@1.8.0-alpha.0...@quiet/identity@1.8.0) (2023-08-28)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [1.8.0-alpha.0](https://github.com/TryQuiet/identity/compare/@quiet/identity@1.7.0...@quiet/identity@1.8.0-alpha.0) (2023-08-25)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
 # [1.7.0](https://github.com/TryQuiet/identity/compare/@quiet/identity@1.6.0...@quiet/identity@1.7.0) (2023-08-17)
 
 **Note:** Version bump only for package @quiet/identity